+++ vaadin-maps-leaflet-flow-demo/src/main/java/software/xdev/vaadin/data/entity/LPointConverter.java
@@ -1,4 +1,4 @@
-package software.xdev.vaadin.maps.leaflet.flow.data.entity;
+package software.xdev.vaadin.data.entity;
 
 import java.util.ArrayList;
 import java.util.List;
