+++ vaadin-maps-leaflet-flow-demo/src/main/java/software/xdev/vaadin/data/entity/Rectangle.java
@@ -1,10 +1,6 @@
-package software.xdev.vaadin.maps.leaflet.flow.data.entity;
-
-import java.util.ArrayList;
-import java.util.List;
+package software.xdev.vaadin.data.entity;
 
 import jakarta.persistence.Convert;
-import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 import software.xdev.vaadin.maps.leaflet.flow.data.LPoint;
