+++ vaadin-maps-leaflet-flow-demo/src/main/java/software/xdev/vaadin/data/repository/MarkerRepository.java
@@ -1,0 +1,10 @@
+package software.xdev.vaadin.data.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import software.xdev.vaadin.data.entity.Marker;
+
+
+
+public interface MarkerRepository extends JpaRepository<Marker, Long> {
+}
