+++ vaadin-maps-leaflet-flow-demo/src/main/java/software/xdev/vaadin/data/repository/PolylineRepository.java
@@ -1,8 +1,8 @@
-package software.xdev.vaadin.maps.leaflet.flow.data.repository;
+package software.xdev.vaadin.data.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import software.xdev.vaadin.maps.leaflet.flow.data.entity.Polyline;
+import software.xdev.vaadin.data.entity.Polyline;
 
 
 public interface PolylineRepository extends JpaRepository<Polyline, Long>
